# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/history/create": {
            "post": {
                "tags": [
                    "History"
                ],
                "summary": "Create history",
                "description": "Create new budget history",
                "operationId": "History",
                "requestBody": {
                    "description": "Pass history credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount",
                                    "description",
                                    "budget_type"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "format": "id",
                                        "example": "flutW0MZyjQikx7mPpslUq37ryl1"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "12345"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "₦20,000 was added to Education budget plan"
                                    },
                                    "budget_type": {
                                        "type": "string",
                                        "example": "Education"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong user details. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "History Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/histories/{user_id}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get histories",
                "description": "Get all budget histories by user",
                "operationId": "GetHistories",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong user details. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "data objects"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/histories/{user_id}/{budget_type}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get single history",
                "description": "Get all budget of a particular type histories belonging to a particular user user",
                "operationId": "GetHistory",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong user details. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "data objects"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}